refactor: update breadcrumbs code, use new modular breadcrumb db call to query newbie post thread breadcrumbs
