+++ modules/ept-breadcrumbs/db/db.js
@@ -1,0 +1,3 @@
+var path = require('path');
+var dbc = require(path.normalize(__dirname + '/../../../db.js'));
+module.exports = dbc;
