+++ modules/ept-breadcrumbs/db/index.js
@@ -1,0 +1,5 @@
+var path = require('path');
+
+module.exports = {
+  getBreadcrumbs: require(path.normalize(__dirname + '/getBreadcrumbs')),
+};
